lab/llama_agent/requirements.txt
langchain-groq
langchain-core
langchain
python-dotenv
tenacity
streamlit
